requirements.txt
coloredlogs
Pyrogram>=2.0
PyYAML
TgCrypto
